+++ src/main/java/org/bytedeco/javacv/LeptonicaFrameConverter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2019 Samuel Audet
+ * Copyright (C) 2018-2021 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -26,7 +26,6 @@ import java.nio.Buffer;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import org.bytedeco.javacpp.BytePointer;
-import org.bytedeco.javacpp.IntPointer;
 import org.bytedeco.javacpp.Loader;
 import org.bytedeco.javacpp.Pointer;
 
@@ -44,6 +43,7 @@ public class LeptonicaFrameConverter extends FrameConverter<PIX> {
     static { Loader.load(org.bytedeco.leptonica.global.lept.class); }
 
     PIX pix;
+    BytePointer frameData, pixData;
     ByteBuffer frameBuffer, pixBuffer;
 
     static boolean isEqual(Frame frame, PIX pix) {
@@ -63,13 +63,22 @@ public class LeptonicaFrameConverter extends FrameConverter<PIX> {
         } else if (!isEqual(frame, pix)) {
             Pointer data;
             if (ByteOrder.nativeOrder().equals(ByteOrder.LITTLE_ENDIAN)) {
-                data = new BytePointer(frame.imageHeight * frame.imageStride);
+                if (pixData == null || pixData.capacity() < frame.imageHeight * frame.imageStride) {
+                    if (pixData != null) {
+                        pixData.releaseReference();
+                    }
+                    pixData = new BytePointer(frame.imageHeight * frame.imageStride).retainReference();
+                }
+                data = pixData;
                 pixBuffer = data.asByteBuffer().order(ByteOrder.BIG_ENDIAN);
             } else {
                 data = new Pointer(frame.image[0].position(0));
             }
+            if (pix != null) {
+                pix.releaseReference();
+            }
             pix = PIX.create(frame.imageWidth, frame.imageHeight, frame.imageChannels * 8, data)
-                     .wpl(frame.imageStride / 4 * Math.abs(frame.imageDepth) / 8);
+                     .wpl(frame.imageStride / 4 * Math.abs(frame.imageDepth) / 8).retainReference();
         }
 
         if (ByteOrder.nativeOrder().equals(ByteOrder.LITTLE_ENDIAN)) {
@@ -111,12 +120,23 @@ public class LeptonicaFrameConverter extends FrameConverter<PIX> {
             frame.imageChannels = pix.d() / 8;
             frame.imageStride = pix.wpl() * 4;
             if (ByteOrder.nativeOrder().equals(ByteOrder.LITTLE_ENDIAN)) {
-                Pointer data = new BytePointer(frame.imageHeight * frame.imageStride);
-                frameBuffer = data.asByteBuffer().order(ByteOrder.LITTLE_ENDIAN);
-                frame.opaque = data;
+                if (frameData == null || frameData.capacity() < frame.imageHeight * frame.imageStride) {
+                    if (frameData != null) {
+                        frameData.releaseReference();
+                    }
+                    frameData = new BytePointer(frame.imageHeight * frame.imageStride).retainReference();
+                }
+                frameBuffer = frameData.asByteBuffer().order(ByteOrder.LITTLE_ENDIAN);
+                frame.opaque = frameData;
                 frame.image = new Buffer[] { frameBuffer };
             } else {
-                frame.opaque = tempPix != null ? pix.clone() : pix;
+                if (tempPix != null) {
+                    if (this.pix != null) {
+                        this.pix.releaseReference();
+                    }
+                    this.pix = pix = pix.clone();
+                }
+                frame.opaque = pix;
                 frame.image = new Buffer[] { pix.createBuffer() };
             }
         }
@@ -130,5 +150,21 @@ public class LeptonicaFrameConverter extends FrameConverter<PIX> {
             pixDestroy(tempPix);
         }
         return frame;
+    }
+
+    @Override public void close() {
+        super.close();
+        if (pix != null) {
+            pix.releaseReference();
+            pix = null;
+        }
+        if (pixData != null) {
+            pixData.releaseReference();
+            pixData = null;
+        }
+        if (frameData != null) {
+            frameData.releaseReference();
+            frameData = null;
+        }
     }
 }